+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# NEXT VERSION
+
+- Feature: add "monthUtils" with functions to retrieve the months names in the user's locale
+- Feature: add "weekUtils" with functions to retrieve the week days' names in the user's locale
+
 # 1.1.1
 
 - Fix: "numberFormatter" missing from library